chore: refactor driver module
This change renames the top-level driver module to "io" and moves the driver and Op to a driver module within the runtime module.
