+++ src/fs/file.rs
@@ -1,7 +1,9 @@
 use crate::buf::{BoundedBuf, BoundedBufMut, IoBuf, IoBufMut, Slice};
 use crate::driver::{Op, SharedFd};
 use crate::fs::OpenOptions;
+use crate::io::SharedFd;
 
+use crate::runtime::driver::op::Op;
 use std::fmt;
 use std::io;
 use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd, RawFd};
