+++ src/fs/open_options.rs
@@ -1,6 +1,6 @@
-use crate::driver::Op;
 use crate::fs::File;
 
+use crate::runtime::driver::op::Op;
 use std::io;
 use std::os::unix::fs::OpenOptionsExt;
 use std::path::Path;
