+++ src/io/open.rs
@@ -1,7 +1,7 @@
-use crate::driver::{self, Op, SharedFd};
 use crate::fs::{File, OpenOptions};
+use crate::io::SharedFd;
 
-use crate::driver::op::{self, Completable};
+use crate::runtime::driver::op::{Completable, CqeResult, Op};
 use std::ffi::CString;
 use std::io;
 use std::path::Path;
@@ -17,7 +17,7 @@ impl Op<Open> {
     /// Submit a request to open a file.
     pub(crate) fn open(path: &Path, options: &OpenOptions) -> io::Result<Op<Open>> {
         use io_uring::{opcode, types};
-        let path = driver::util::cstr(path)?;
+        let path = super::util::cstr(path)?;
         let flags = libc::O_CLOEXEC
             | options.access_mode()?
             | options.creation_mode()?
@@ -40,7 +40,7 @@ impl Op<Open> {
 impl Completable for Open {
     type Output = io::Result<File>;
 
-    fn complete(self, cqe: op::CqeResult) -> Self::Output {
+    fn complete(self, cqe: CqeResult) -> Self::Output {
         Ok(File::from_shared_fd(SharedFd::new(cqe.result? as _)))
     }
 }
