+++ src/io/read.rs
@@ -1,8 +1,8 @@
 use crate::buf::BoundedBufMut;
-use crate::driver::{Op, SharedFd};
+use crate::io::SharedFd;
 use crate::BufResult;
 
-use crate::driver::op::{self, Completable};
+use crate::runtime::driver::op::{Completable, CqeResult, Op};
 use std::io;
 
 pub(crate) struct Read<T> {
@@ -42,7 +42,7 @@ where
 {
     type Output = BufResult<usize, T>;
 
-    fn complete(self, cqe: op::CqeResult) -> Self::Output {
+    fn complete(self, cqe: CqeResult) -> Self::Output {
         // Convert the operation result to `usize`
         let res = cqe.result.map(|v| v as usize);
         // Recover the buffer
