+++ src/net/unix/stream.rs
@@ -1,6 +1,6 @@
 use crate::{
     buf::{BoundedBuf, BoundedBufMut, IoBuf},
-    driver::{SharedFd, Socket},
+    io::{SharedFd, Socket},
 };
 use socket2::SockAddr;
 use std::{
